Title: 1, 2, 5 더하기, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 250217/1, 2, 5 더하기/1-2-5-plus.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int dp[1001];
+long long dp[1001];
 int n;
 
 int main() {
